Create Ingredient Form
- Create an ingredient form for admins to add new ingredients.
- Modified DishCreation to use Tabs to switch between dish and ingredient form.
- [BUG] Update featured dishes' images by using the right property.
- [UI-FIX] Add more text options for Typewriter.
- [UI-FIX] Include username beside the profile icon.
- [UI-FIX] Update Tabs in Profile page to use the same variant as the dish-creation page.

#159 #145
