+++ frontend/src/components/pages/Dashboard.jsx
@@ -159,7 +159,7 @@ export default function Dashboard () {
 									>
 										<Image
 											onClick={() => setSelectedDish(dish)}
-											src={dish.photos ? dish.photos[0].getUrl() : '/images/1.jpg'}
+											src={dish.imageUrl || '/images/1.jpg'}
 											fluid
 											style={{
 												width: '100%',
